[FIX] copy sequences locally on left sequence extensions
This PR changes the way how to compute the local alignment on the left
of the current seed. Before this change the source sequence was reversed
in-place, after this change the source sequence is left untouched and
the sequence will be copied.

Local copy needs to include one more character.

Remove _reverseLeftExtension

We don't need this any more as we are not modifying the underlying
sequence any more and use a local copy instead.